+++ src/helix/webhooks/topics/streams.rs
@@ -1,7 +1,0 @@
-//! Webhook topics for streams
-use crate::helix::webhooks::Topic;
-
-pub mod stream_changed;
-
-#[doc(inline)]
-pub use stream_changed::StreamChangedTopic;