+++ src/helix/webhooks/topics/subscriptions.rs
@@ -1,7 +1,0 @@
-//! Webhook topics for subscriptions
-use crate::helix::webhooks::Topic;
-
-pub mod subscription_events;
-
-#[doc(inline)]
-pub use subscription_events::SubscriptionEventsTopic;