+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 
 * Removed enum variants for a lot of error states in helix endpoint responses. Most of these are returned by `HelixRequest_Error::Error`
 * Removed deprecated follow endpoints `Create Users Follows` and `Delete Users Follows`
+* Removed helix `webhook`s and `HelixRequestGetError::InvalidUri`
 
 ## [v0.5.0] - 2021-05-08